move the utilities from __init__ to a new file
